+++ index.js
@@ -34,10 +34,8 @@ class LoginClient {
   getUserData() {
     const cookies = cookieManager.parse(document.cookie);
     const userData = cookies.user && jwtManager.decode(cookies.user);
-    return userData && {
-      userId: userData.sub,
-      email: userData.email
-    };
+    userData.userId = userData.sub;
+    return userData;
   }
 
   /**