+++ src/index.js
@@ -1,8 +1,8 @@
 const cookieManager = require('cookie');
 const base64url = require('base64url');
 
-const HttpClient = require('./src/httpClient');
-const jwtManager = require('./src/jwtManager');
+const HttpClient = require('./httpClient');
+const jwtManager = require('./jwtManager');
 
 let userSessionResolver;
 let userSessionPromise = new Promise(resolve => userSessionResolver = resolve);