+++ CHANGELOG.md
@@ -5,3 +5,4 @@ This is the changelog for [Authress Login](readme.md).
 * Added `authorization-code` exchange specifying necessary parameters `responseMode = query` and `responseType = code`.
 * Calling `authenticate` no longer redirects user to the `redirectUrl` if the user is already logged in. Redirects will still happen for users not logged in. This fixes a problem where the mechanism for redirects did not match the framework redirect mechanism. In some cases causing refresh loops.
 * Return the full OIDC identity when calling `getUserData()` instead of only a limited set.
+* Add `tenantLookupIdentifier` property to `authenticate` to specify a specific tenant to use rather than just the `connection`.